完全重构，将数据库和模型完全合并。去除了不需要的数据库迁移功能。
