--- app/database/__init__.py
@@ -1,1 +1,0 @@
-from .models import Department, Leave, Report, SignSheet, User, WorkArrangement
