+++ app/model/WorkArrangement.py
@@ -1,17 +1,36 @@
-class WorkArrangement:
-    # employeeID员工工号 int
-    # employeeName员工姓名 string
-    # workDate工作日期 string
-    # assignment工作任务 string
+class WorkArrangement(db.Model):  # 工作安排
+    __tablename__ = 'arrangement'
+    arragementID = db.Column(db.Integer, primary_key=True)
+    staffID = db.Column(db.Integer, db.ForeignKey('user.ID'))  # 员工标号
+    departmentID = db.Column(
+        db.Integer, db.ForeignKey('department.ID'))  # 部门标号
+    workDate = db.Column(db.Date, nullable=False)  # 工作日期
+    beginTime = db.Column(db.Time, nullable=False)  # 开始时间
+    endTime = db.Column(db.Time, nullable=False)  # 结束时间
+    assignment = db.Column(db.String(50))  # 工作安排
 
-    def __init__(self, employeeID, employeeName, workDate, assignment):
-        self.employeeID = employeeID
-        self.employeeName = employeeName
-        self.workDate = workDate
-        self.assignment = assignment
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+
+    def __repr__(self):
+        return '<WorkArrangement %i %i>' % (self.staffID, self.departmentID)
+
+    @classmethod
+    def findAll():
+        return WorkArrangement.query.all()
+
+    @classmethod
+    def getInfoByID(self, arragementID):
+        return WorkArrangement.query.get(arragementID)
+
+    @classmethod
+    def getInfoBystaffID(self, staffID):
+        return WorkArrangement.query.filter_by(staffID=staffID).all()
+
+    @classmethod
+    def getInfoBydepID(self, departmentID):
+        return WorkArrangement.query.filter_by(departmentID=departmentID).all()
 
     def show_work_arrangement(self):
-        """show_work_arrangement展示工作安排"""
+        """展示工作安排"""
         pass
-
-
